Improve stats datastore counting
We were using this check:

  resource[u'url_type'] in [u'datastore', u'upload', u'dataset']

which had to be passed before we'd ask the datastore what the count
on the resource was, however if you don't upload the data (for example
you link to an external URL source) then this fails even though the
datastore will have ingested the data. This has been removed in favour
of checking the datastore_active property of the resource.

Secondly, this commit just tidies up some of the logic, most notably
we don't just get the first 200 packages, we get all of them,
iteratively.