Minor formatting changes and account for duplicate downloads
The gbif api returns the download events in reverse chronological order so
the first record we receive is the latest download. This is an issue when
iterating over the data as we could, if new downloads occur between page
requests, receive data about a download twice. Using a set of seen dois
resolves this issue.