refactor: remove dependency on ckanext-ckanpackager by copying its model
The only part of the ckanext-ckanpackager this extension used was the model so now that the ckanext-ckanpackager extensions is deprecated, just copy the model over to this extension to enable us to remove the deprecated dependency. In the long run we'll tidy this up but this works as a short term fix.

Closes: #65